requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
humanfriendly==10.0
idna==3.10
influxdb==5.3.2
influxdb-client==1.48.0
msgpack==1.1.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
reactivex==4.0.4
requests==2.32.3
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
